refactor: [5단계-자동차 경주(리팩터링)] (1)MVC 모델에 맞춰 패키지 분리 (2)CarGroup에서 입력받은 이름을 구분자 컴마에 맞춰 나누던 책임을 InputView로 옮김
