+++ src/main/java/racingcar/domain/Car.java
@@ -1,7 +1,7 @@
-package racingcar;
+package racingcar.domain;
 
 public class Car {
-    private Name name;
+    private final Name name;
     private Position position;
 
     public Car(String carName) {
