+++ src/main/java/racingcar/domain/RacingMain.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.controller.Racing;
 
 public class RacingMain {
     public static void main(String[] args) {
